Add search and metadata update for matches (#2)
The match search by string is fairly limited. It will attempt to find a match by finding the event it took place on. It does this by looking for a promotion and a date, then trying to find an event by that promotion that took place on that date. It will then list all the matches from that show, and use the match portion of the search string if present to try and rank the correct match to the top.

This will likely fail if promotions or dates are not included, or if the date doesn't match the recording date of the event. I'd like to improve this in the future, but haven't come up with a better solution yet.

Supports searching by a search string of format `{promotion} - {date} - M - {match}` or by an ID of format `cm-id:xxxxxx:y`, where `xxxxxx` is an event page id and `y` is the index (from 1) of the match on the card
Fills in studio and date from event info table
Fills in summary using information from the event info table
Sets the CAGEMATCH matchguide rating for a match if present
Sets Plex roles for each worker in the All Workers list whose string is contained in the title string
Adds match to collections if configured that way

Unrelated fixes:
Don't set a studio if the promotion field is one of the freelance categories
Don't include a promotion in summaries if the promotion field is one of the freelance categories
Make dictionary key queries safer