Change scoring algorithm for matches
When using the default scoring method for matches, the results were
often incorrect. By default, fuzzywuzzy uses a weighted combination
of scorers. Changing this method to `token_set_ratio` seems to
improve the order of results. From reading this blog post by the
creators of fuzzywuzzy, it seems like the Token Sort and
Token Set scorers are intended for almost this exact scenario:
(https://chairnerd.seatgeek.com/fuzzywuzzy-fuzzy-string-matching-in-python)