Use filename for search if present
Plex's scanner will often not give you the entire filename in
`media.name` to search with. We can't fix that without writing a
scanner, but for automatic searches, `media` also contains `filename`,
an encoded path to the file. From this, we can get the raw filename,
and use this to parse instead. This should improve some of the
matching as well, as dashes and other info will be preserved.

Manual searches can use the raw search text (`media.name`) as previous.