Add option to remove promotion intials from events
Adds a config option allowing promotion abbreviations to be removed
from titles. By default, this is only done when events are added to
collections for the promotion. This can also be done for every event
regardless of collection or disabled entirely.

Also adds some more test files, and changes some logging of GET
requests.