Allow for optional whitespace in cm-id search regex
When doing a copy/paste of a CM ID from the web the value will somtimes
paste into the Plex matcher UI with a space between ‘cm-id:’ and the
ID itself.

This should allow for that as an optional part of the regex match and
include it in the “cm-id” part of the regex.